# src/main/resources/swagger/v1/hotelaria-api.yml
swagger: '2.0'
info:
  version: 1.0.0
  title: Lorie API

host: localhost:8080
basePath: /
tags:
  - name: hoteis
  - name: hospedes
  - name: reservas
schemes:
  - http
  - https

paths:
  '/hoteis':
    get:
      tags:
        - hoteis
      summary: Listagem de hoteis
      description: Lista todos hoteis cadastrados na rede de hotelaria
      operationId: retrieveHotels
      produces:
        - application/json
      responses:
        '200':
          description: Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/HotelResponse'

    post:
      tags:
        - hoteis
      summary: Create a new hotel
      description: Create a new hotel
      operationId: createHotel
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Hotel object that needs to be added
          required: true
          schema:
            $ref: '#/definitions/HotelRequest'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/HotelResponse'
  '/hoteis/{hotelId}':
    get:
      tags:
        - hoteis
      summary: Retorna um hotel
      description: Retorna um hotel com o ID informado
      operationId: retrieveHotelById
      produces:
        - application/json
      parameters:
        - name: hotelId
          in: path
          description: ID of hotel to return
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/HotelResponse'
        '404':
          description: Not Found

    put:
      tags:
        - hoteis
      summary: Update an existing hotel
      description: Update an existing hotel
      operationId: updateHotel
      produces:
        - application/json
      parameters:
        - name: hotelId
          in: path
          description: ID of hotel to return
          required: true
          type: integer
          format: int64
        - in: body
          name: body
          description: Hotel object that needs to be updated
          required: true
          schema:
            $ref: '#/definitions/UpdateHotelRequest'
      responses:
        '200':
          description: Ok
        '404':
          description: Not Found

    delete:
      tags:
        - hoteis
      summary: Remove um hotel
      description: Remove um hotel com o ID informado
      operationId: removeHotel
      produces:
        - application/json
      parameters:
        - name: hotelId
          in: path
          description: ID of hotel to return
          required: true
          type: integer
          format: int64
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
  '/hoteis/{hotelId}/acomodacoes-disponiveis':
    get:
      tags:
        - hoteis
      summary: Listagem de acomodações disponíveis
      description: Lista todas acomodações disponíveis em um hotel
      operationId: retrieveDisponibilidade
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of hotel to return
          required: true
          type: integer
          format: int64
        - name: from
          in: query
          description: Data de checkin
          required: true
          type: string
        - name: to
          in: query
          description: Data de checkout
          required: true
          type: string
      responses:
        '200':
          description: Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/DisponibilidadeResponse'

  '/hospedes':
    get:
      tags:
        - hospedes
      summary: Listagem de hospedes
      description: Lista todos hospedes cadastrados na rede de hotelaria
      operationId: retrieveHospedes
      produces:
        - application/json
      responses:
        '200':
          description: Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/HospedeResponse'

    post:
      tags:
        - hospedes
      summary: Criação de um novo hospede
      description: Criação de um novo hospede com os dados informados
      operationId: createHospede
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Hospede object that needs to be added
          required: true
          schema:
            $ref: '#/definitions/HospedeRequest'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/HospedeResponse'
  '/hospedes/{hospedeId}':
    get:
      tags:
        - hospedes
      summary: Retorna um hospede
      description: Retorna um hospede com o ID informado
      operationId: retrieveHospedeById
      produces:
        - application/json
      parameters:
        - name: hospedeId
          in: path
          description: ID of hospede to return
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/HospedeResponse'
        '404':
          description: Not Found

    put:
      tags:
        - hospedes
      summary: Atualiza um hospede
      description: Atualiza um hospede com os dados informados
      operationId: updateHospede
      produces:
        - application/json
      parameters:
        - name: hospedeId
          in: path
          description: ID of hospede to return
          required: true
          type: integer
          format: int64
        - in: body
          name: body
          description: Hospede object that needs to be updated
          required: true
          schema:
            $ref: '#/definitions/HospedeRequest'
      responses:
        '200':
          description: Ok
        '404':
          description: Not Found

    delete:
      tags:
        - hospedes
      summary: Remove um hospede
      description: Remove um hospede com o ID informado
      operationId: removeHospede
      produces:
        - application/json
      parameters:
        - name: hospedeId
          in: path
          description: Identificador do hospede
          required: true
          type: integer
          format: int64
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
  '/hospedes/{hospedeId}/reservas':
    get:
      tags:
        - reservas
      summary: Listagem de reservas
      description: Lista todas reservas de um hospede
      operationId: retrieveReservas
      produces:
        - application/json
      parameters:
        - name: hospedeId
          in: path
          description: ID of hospede to return
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/ReservaAcomodacaoResponse'
    post:
      tags:
        - reservas
      summary: Solicitação de reserva
      description: Solicitação de reserva de acomodação
      operationId: makeReservation
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Solicitacao de reserva
          required: true
          schema:
            $ref: '#/definitions/SolicitacaoReservaRequest'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/ReservaAcomodacaoResponse'

definitions:
  HotelRequest:
    type: object
    properties:
      nome_fantasia:
        type: string
        example: 'New Star Hotel'
      setor:
        type: string
        example: 'FAMILIA'
        enum:
          - 'FAMILIA'
          - 'ADULTOS'
          - 'PERMITE_PETS'
      tamanho:
        type: integer
        format: int64
      is_central:
        type: boolean
      categoria:
        type: string
        example: 'TRADICIONAL'
        enum:
          - 'TRADICIONAL'
          - 'POUSADA'
          - 'APART_HOTEL'
          - 'CONDOMINIO_DE_HOTEIS'
          - 'RESORT'
      acomodacoes:
        type: array
        items:
          $ref: '#/definitions/AcomodacaoRequest'
      registro_imobiliario:
        $ref: '#/definitions/RegistroImobiliarioRequest'
    required:
      - nome_fantasia
      - setor
      - tamanho
      - is_central
      - categoria
      - acomodacoes
      - registro_imobiliario

  UpdateHotelRequest:
    type: object
    properties:
      nome_fantasia:
        type: string
        example: 'New Star Hotel'
      setor:
        type: string
        example: 'FAMILIA'
        enum:
          - 'FAMILIA'
          - 'ADULTOS'
          - 'PERMITE_PETS'
      tamanho:
        type: integer
        format: int64
      is_central:
        type: boolean
      categoria:
        type: string
        example: 'TRADICIONAL'
        enum:
          - 'TRADICIONAL'
          - 'POUSADA'
          - 'APART_HOTEL'
          - 'CONDOMINIO_DE_HOTEIS'
          - 'RESORT'

  HotelResponse:
    type: object
    properties:
      id:
        type: integer
        format: int64
      nome_fantasia:
        type: string
      setor:
        type: string
      tamanho:
        type: integer
        format: int64
      is_central:
        type: boolean
      categoria:
        type: string
      acomodacoes:
        type: array
        items:
          $ref: '#/definitions/AcomodacaoResponse'
      registro_imobiliario:
        $ref: '#/definitions/RegistroImobiliarioResponse'

  RegistroImobiliarioRequest:
    type: object
    properties:
      data_registro:
        type: string
        format: date
      numero_registro_prefeitura:
        type: string
        example: 'NP123456'
      georreferenciamento:
        $ref: '#/definitions/GeorreferenciamentoRequest'
    required:
      - data_registro
      - numero_registro_prefeitura
      - georreferenciamento

  RegistroImobiliarioResponse:
    type: object
    properties:
      id:
        type: integer
        format: int64
      data_registro:
        type: string
      numero_registro_prefeitura:
        type: string
      georreferenciamento:
        $ref: '#/definitions/GeorreferenciamentoResponse'

  GeorreferenciamentoRequest:
    type: object
    properties:
      latitude:
        type: number
        format: double
      longitude:
        type: number
        format: double
    required:
      - latitude
      - longitude

  GeorreferenciamentoResponse:
    type: object
    properties:
      id:
        type: integer
        format: int64
      latitude:
        type: string
      longitude:
        type: string

  PoliticaDeUsoRequest:
    type: object
    properties:
      nome:
        type: string
      descricao:
        type: string
    required:
      - nome
      - descricao

  PoliticaDeUsoResponse:
    type: object
    properties:
      id:
        type: integer
        format: int64
      nome:
        type: string
      descricao:
        type: string

  AcomodacaoRequest:
    type: object
    properties:
      numero:
        type: integer
        format: int64
      valor_diaria:
        type: number
        format: double
      tipo:
        type: string
        example: 'STANDARD'
        enum:
          - 'STANDARD'
          - 'STANDARD_COM_VISTA'
          - 'DELUXE'
          - 'DELUXE_COM_VISTA'
          - 'SUITE'
          - 'SUITE_COM_VISTA'
          - 'EXECUTIVE'
          - 'EXECUTIVE_COM_VISTA'
          - 'PRESIDENTIAL_SUITE'
          - 'ROYAL_SUITE'
      capacidade:
        type: integer
        format: int32
      politicas_de_uso:
        type: array
        items:
          type: integer
          format: int64

  AcomodacaoResponse:
    type: object
    properties:
      numero:
        type: integer
        format: int64
      valor_diaria:
        type: number
        format: double
      tipo:
        type: string
      capacidade:
        type: integer
        format: int32
      politicas_de_uso:
        type: array
        items:
          $ref: '#/definitions/PoliticaDeUsoResponse'

  ReservaAcomodacaoResponse:
    type: object
    properties:
      acomodacao:
        $ref: '#/definitions/AcomodacaoResponse'
      data_esperada_checkin:
        type: string
      data_esperada_checkout:
        type: string
      data_checkin:
        type: string
      data_checkout:
        type: string
      valor_total:
        type: string

  SolicitacaoReservaRequest:
    type: object
    properties:
      hospede_id:
        type: integer
        format: int64
      numero_acomodacao:
        type: integer
        format: int64
      hotel_id:
        type: integer
        format: int64
      data_checkin:
        type: string
      data_checkout:
        type: string
    required:
      - hospede_id
      - numero_acomodacao
      - hotel_id
      - data_esperada_check_in
      - data_esperada_check_out
      
  DisponibilidadeResponse:
    type: object
    properties:
      data:
        type: string
      quantidade_disponivel:
        type: integer
        format: int32
      acomodacoes_disponiveis:
        type: array
        items:
          $ref: '#/definitions/AcomodacaoResponse'

  HospedeRequest:
    type: object
    properties:
      nome:
        type: string
        example: 'João da Silva'
      sexo:
        type: string
        example: 'MASCULINO'
        enum:
          - 'MASCULINO'
          - 'FEMININO'
      email:
        type: string
        example: 'joao@gmail.com'
      celular:
        type: string
        example: '11999999999'
      data_nascimento:
        type: string
        format: date
      documento_identificacao:
        $ref: '#/definitions/DocumentoIdentificacaoRequest'
      data_registro:
        type: string
        format: date
    required:
      - nome
      - sexo
      - email
      - celular
      - data_nascimento
      - documento_identificacao
      - data_registro

  HospedeResponse:
    type: object
    properties:
      id:
        type: integer
        format: int64
      nome:
        type: string
      sexo:
        type: string
      email:
        type: string
      celular:
        type: string
      data_nascimento:
        type: string
      documento_identificacao:
        $ref: '#/definitions/DocumentoIdentificacaoResponse'
      data_registro:
        type: string

  PessoaRequest:
    type: object
    properties:
      nome:
        type: string
        example: 'João da Silva'
      sexo:
        type: string
        example: 'MASCULINO'
        enum:
          - 'MASCULINO'
          - 'FEMININO'
      email:
        type: string
        example: 'joao@gmail.com'
      celular:
        type: string
        example: '11999999999'
      data_nascimento:
        type: string
        format: date
      documento_identificacao:
        $ref: '#/definitions/DocumentoIdentificacaoRequest'

  PessoaResponse:
    type: object
    properties:
      id:
        type: integer
        format: int64
      nome:
        type: string
      sexo:
        type: string
      email:
        type: string
      celular:
        type: string
      data_nascimento:
        type: string
        format: date
      documento_identificacao:
        $ref: '#/definitions/DocumentoIdentificacaoResponse'

  DocumentoIdentificacaoRequest:
    type: object
    properties:
      tipo:
        type: string
        example: 'CPF'
        enum:
          - 'CPF'
          - 'CNPJ'
          - 'RG'
          - 'PASSAPORTE'
      numero:
        type: string
        example: '12345678901'
    required:
      - tipo
      - numero

  DocumentoIdentificacaoResponse:
    type: object
    properties:
      id:
        type: integer
        format: int64
      tipo:
        type: string
      numero:
        type: string